Update RFC0001-Mutual-Exclusion-Parameters-and-Properties.md
Based on the feedback, this proposal as it's written is not ready for the next stage.  Any RFC attempting to resolve this should consider the authoring experience, code review experience (how easy it is to understand intent from the code, particularly for complex examples like invoke-command), and also considerations for improving readability of help.